Moved copy-coordinates
