+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/copy-coordinates/index.tsx
